Missed some directory renames
